Add dto classes to display information for categories and questions, make service and controller methods to get by id, get all, delete and create elements

+++ src/main/java/com/quizapp/questions/service/CategoryServiceImpl.java
@@ -1,13 +1,16 @@
 package com.quizapp.questions.service;
 
 import com.quizapp.questions.model.dto.AddCategoryDTO;
+import com.quizapp.questions.model.dto.CategoryDTO;
 import com.quizapp.questions.model.entity.Category;
 import com.quizapp.questions.repository.CategoryRepository;
 import com.quizapp.questions.service.interfaces.CategoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -18,6 +21,28 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public Optional<Category> findCategoryByName(String name) {
         return this.categoryRepository.findByName(name);
+    }
+
+    @Override
+    public List<CategoryDTO> getAllCategories() {
+        return this.categoryRepository.findAll().stream()
+                .map(this::categoryToDTO)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public CategoryDTO getCategoryById(Long id) {
+        return this.categoryRepository.findById(id)
+                .map(this::categoryToDTO)
+                .orElse(null);
+    }
+
+    private CategoryDTO categoryToDTO(Category category) {
+        return CategoryDTO.builder()
+                .id(category.getId())
+                .name(category.getName())
+                .description(category.getDescription())
+                .build();
     }
 
     @Override
@@ -32,7 +57,12 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void deleteCategoryById(Long id) {
+    public boolean deleteCategoryById(Long id) {
+        if (!this.categoryRepository.existsById(id)) {
+            return false;
+        }
+
         this.categoryRepository.deleteById(id);
+        return true;
     }
 }
